Update chapter07.md
Some background on the proposed changes:

Having three elements in an example always seems skimpy or stingy to me. Why would you go to all this effort, for example, to sort just three things? I like five as a minimum. Feel free to ignore this change. ;-) 

Best to be consistent and stick to 'you' and avoid 'we' or 'one'.

Point to Wikipedia definition articles wherever possible.

'However' is best place inside a sentence - where it implies a 'but' and not at the front of a sentence where it implies 'and'.

Add a bibliography - point to sources - that may help you learn even more.